feat: migrate banner to tsx (#909)
